Added maiapy, New Part and Measure methods and improve build system

+++ scripts/make-clean.py
@@ -5,20 +5,29 @@ from terminal_colors import *
 numArgs = len(sys.argv)
 
 if (numArgs != 2):
-    print(f"{color.FAIL}[ERROR] You MUST pass 1 argument: 'all', 'static', 'shared' or 'module'!{color.ENDC}")
+    print(f"{color.FAIL}[ERROR] You MUST pass 1 argument: 'all', 'static', 'shared', 'module' or 'dist'!{color.ENDC}")
 
-libType = sys.argv[0]
+cleanOption = sys.argv[1]
 
 # ===== VALIDATE INPUT ARGUMENTS ===== #
 
-# At this moment, we are not using the 'libType' input argument
-# In the future, we will implement a 'clear' for each lib type folder
 
-print(f"{color.OKGREEN}Deleting 'build' and 'dist' folders...{color.ENDC}", end = "")
+if (cleanOption == "all"):
+    print(f"{color.OKGREEN}Deleting 'build' and 'dist' folders...{color.ENDC}", end = "")
 
-rmtree("./build", True)
-rmtree("./dist", True)
-rmtree("__pycache__", True)
-rmtree("tests/__pycache__", True)
+    rmtree("./build", True)
+    rmtree("./dist", True)
+    rmtree("__pycache__", True)
+    rmtree("tests/__pycache__", True)
+    rmtree("maiapy/__pycache__", True)
+    rmtree("scripts/__pycache__", True)
+    print(f"{color.OKGREEN}Done!{color.ENDC}")
+    sys.exit()
 
-print(f"{color.OKGREEN}Done!{color.ENDC}")
+if (cleanOption == "dist"):
+    print(f"{color.OKGREEN}Deleting 'dist' folder...{color.ENDC}", end = "")
+    rmtree("./dist", True)
+    print(f"{color.OKGREEN}Done!{color.ENDC}")
+    sys.exit()
+
+
